+++ packages/System/DTO/WordPressInformation.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace WPShop\System\DTO;
+
+final readonly class WordPressInformation
+{
+    public function __construct(
+        public string $version,
+        public string $locale,
+        public string $timezone,
+        public bool $multisite,
+        public bool $debug
+    ) {
+    }
+}